+++ platform/swift/source/LoggerObjc.swift
@@ -381,6 +381,21 @@ public final class LoggerObjc: NSObject {
         )
     }
 
+    /// Logs a critical level message to the default logger instance.
+    ///
+    /// - parameter message: The message to log.
+    /// - parameter fields:  The extra fields to send as part of the log.
+    @objc
+    public static func logCritical(_ message: String, fields: [String: String]?) {
+        Capture.Logger.logCritical(
+            message,
+            file: nil,
+            line: nil,
+            function: nil,
+            fields: fields
+        )
+    }
+
     /// Logs a message at a specified level to the default logger instance.
     ///
     /// - parameter level:   The severity of the log.