+++ platform/swift/source/Logging.swift
@@ -349,6 +349,33 @@ extension Logging {
         )
     }
 
+    /// Logs a critical level message to the default logger instance.
+    ///
+    /// - parameter message:  The message to log.
+    /// - parameter file:     The unique file identifier that has the form module/file.
+    /// - parameter line:     The line number where the log is emitted.
+    /// - parameter function: The name of the function from which the log is emitted.
+    /// - parameter fields:   The extra fields to send with the log.
+    /// - parameter error:    The error to log.
+    public func logCritical(
+        _ message: @autoclosure () -> String,
+        file: String? = #file,
+        line: Int? = #line,
+        function: String? = #function,
+        fields: Fields? = nil,
+        error: Error? = nil
+    ) {
+        self.log(
+            level: .critical,
+            message: message(),
+            file: file,
+            line: line,
+            function: function,
+            fields: fields,
+            error: error
+        )
+    }
+
     /// Logs information about a network request. Provides default values for
     /// `file`, `line`, and `function` parameters.
     ///